// lib/types/drug.ts
export type Drug = {
  rowIndex: number;
  name: string;
  ingredient: string;
  category: string;
  subcategory: string;
  form: string;
  dose: string;
  position: string;
  sameDrug: string;
  crush: string;
  suspension: string;
  saline: string;
  glucose: string;
  caution: string;
  note: string;
};